+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - mark webxdc app context as secure #3413
+- Experimental: Related Chats
 
 ### Changed
 - update deltachat-node and deltachat/jsonrpc-client to `v1.123.0`
@@ -22,6 +23,7 @@
 
 ### Added
 - Show video chat instance URLs as subtitles #3369
+- Add similar chats to group profile #3379
 
 ### Changed
 - Offer to copy non-HTTP links to the clipboard instead of trying to open them in webxdc source code link and inside of html emails.